Move Big Book of Acronyms info to annotations
Kubernetes label values are used internally by Kubernetes, and should
be data that we might want to select based on. Further, they have a
limited character set for values, meaning we can't put `application`,
`owner`, `infrastructure-support`, or `source-code` in them:

> Valid label values must be 63 characters or less and must be empty or
> begin and end with an alphanumeric character ([a-z0-9A-Z]) with dashes
> (-), underscores (_), dots (.), and alphanumerics between.
- https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#syntax-and-character-set

[Annotations](https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations/)
are a much better fit for this data, and the docs even mention this
kind of data:

> - User or tool/system provenance information, such as URLs of related
>   objects from other ecosystem components.
> […]
> - Phone or pager numbers of persons responsible, or directory entries
>   that specify where that information can be found, such as a team web
>   site.

I'm also making sure all the values are wrapped in double-quotes, just
to be sure the data is parsed as I intend.