Introduce index template for audit logs
Elasticsearch automatically generates a mapping for records and indexes all fields. However, the audit logs have a few
elements that may differ between records and that leads to a number of records being rejected. This change introduces
an index template that allows only a subset of safe and useful fields to be indexed.