Remove UAT RDS database
As part of a spike to investigate moving from containerised postgres
to RDS, an RDS database was created for the laa-apply-for-legalaid-uat
namespace. This is no longer needed, as this approach is not suitable
for our approach to UAT.

A separate commit will create RDS instances for staging and production,
where this is a suitable approach.