Update default LimitRange and ResourceQuota values
This updates all the LimitRange and ResourceQuota values for all namespaces that are currently using them (namespaces
that have used the templates but have not customised them further.

The ratio of LimitRange and ResourceQuota for all values is now constant and equal to 24 (ie. with default values, it's
up to 24 containers in a namespace).

This change should address various scheduling issues we've been struggling with and set new defaults that hopefully will
work better for the majority of namespaces we host.