Delete unused c100-application-production live-0 namespace
As part of a previous (old) PR #620, we deleted any AWS resources in the `production` namespace, but we didn't remove it as we were waiting for a migration to happen.

Now we are all good and this Live-0 namespace can be deleted.
We believe we've deleted everything inside the namespace (pods, secrets, ingress, etc.).