Remove ResourceQuota for the logging namespace
Since we deploy a DaemonSet of fluentd agents in this namespace, we need to update the ResourceQuota accordingly, with
any change to the cluster configuration.

The risk is that fluentd will not be able to be scheduled on all nodes and since it is a critical system service, it is
easier to remove the restriction.