Enable reporter to handle more memory formats
If limits/quotas are set to decimal Gi values, 
e.g. 6.4Gi, you end up with a value measures in
bits, with an 'm' suffix.
If a value is set to 0, the str value here will be
"0".

This commit enables memory_value to cope sensibly
with both of these formats