Halve the requests limits for new namespaces
#1012 reduced the per-container resource requests
(i.e. resources the scheduler will *reserve* for
every container in a namespace).

Having reduced those values, we should be able to
reduce the overall resources that will be reserved
for the namespaces as a whole.

This commit reduces those values by half, which is
a conservative change.

Looking at our actual usage, out of 114 namespaces
in live-1, 103 are using less than 2Gi of memory,
and 107 are using less than 100m of CPU.

NB: This only puts a limit on the resources the
cluster will reserve for a namespace. It does not
prevent the namespace from *consuming* more than
this.

Relates to #1012
Relates to https://github.com/ministryofjustice/cloud-platform/issues/1094
Closes https://github.com/ministryofjustice/cloud-platform/issues/1093