Add certificate for already existing DNS ZONE
The production ingress should be the "base" DNS
ZONE URI `laa-fee-calculator.service.justice.gov.uk`.

Note, this does not currently exist in live-0, from
which we are migrating, as it uses the default zone
setup. So no "DNS flapping" risks to be concerned with.

The DNS/HOSTED Zone was already created as part of
laa-fee-calculator-staging setup, so I assume we can
just issue a certificate?!