Remove 'limits' from resource quota template
There seems to be no good reason to have a limits
section here. 

If we do have one, then every workload scheduled
onto the cluster needs to fall within the 
namespace's 'request' and 'limits' values (wrt. 
resources allocated to it). However, the 'limits'
values are only useful at runtime, not at schedule
time. 

So, all the limits section does is create another
way for your workload to not be scheduled. This is
confusing and useless.

By removing it, we simplify the way users have to
think about capacity, when sizing their containers