Reduce default request limit values for namespaces
Allocating 1.5 CPU cores and 3G of memory to every
new namespace, by default, is wasteful. 

These values are sufficient for two-thirds of all
the namespaces we have in the cluster.

This is not enough space for most, if not all, of
our production namespaces. This is by design, 
because it will force users to consider how much
resource they need for their namespaces, and raise
PRs to make the necessary adjustments.