Reduce container default request and limits values
Users can specify whatever resources they want to
allocate to their containers, in their deployments

This change reduces the default resources that
will be set aside for every container in a
namespace. These values are representative of the
containers we actually run in the cluster.

They are deliberately set lower than some
containers will need, so that users have to decide
how much resource to assign in their deployments.