Restore container CPU limit to 1000m
The containers in this namespace need a LOT of CPU when they first start (I've seen usage burst to 988m, then settle down to 4m once everything is initialised).
Reducing the CPU limit caused containers to fail to start successfully. My theory is that they're being throttled for CPU, so startup doesn't complete before the readiness probe is tested, and the cluster thinks the container has failed and restarts it.
With the limit back at 1000m, this doesn't seem to happen.