Enable creating namespace with `make namespace`
This makefile target uses the cloud platform tools
image, so the user does not need to install 
terraform, although they do need to have a 
`moj-cp` AWS profile configured, and docker
installed.

This change, once merged, will be used as the
basis for a rewrite of the user guide article on
creating a new namespace.