Clarify the namespace creation process
This commit updates the descriptions of the
terraform variables which the user is asked to
enter, including clarification of what is/isn't
acceptable, and giving more examples.

Also, after the user has finished, they are shown
a message explaining what to do next, and when
their namespace will be available.