Reduce Apply containers CPU allocation
Runtime CPU usage of each of the Apply-for-legal-aid containers stays
under 10m. So allocating 100m for each was an overkill.
We increased the memory allocation per container as some of our
containers use between 256Mi and 900Mi, while others use less than
100Mi.

We normalized the container CPU and Memory allocation for the three
environments/namespaces: UAT, Staging and Production so we have the
same resources everywhere.