Reduce Apply namespaces CPU allocation
Runtime CPU usage of each of the Apply-for-legal-aid namespaces stays
well under 100m (usually below 20m).
So allocating 4000m to 20000m for each namespace was an overkill.

The memory currently used by namespace is bellow 3Gi, and our containers
schedule memory is less than 5Gi, so setting an scheduled allocation of
 8Gi gives us space for even adding a new replica into the namespace.

 We normalized the namespace CPU and Memory allocation for staging and
 production.
 UAT though, as different branches get deployed in the same UAT
 namespace as different releases living together, has higher demands for
 CPU and Memory than the rest, especially when different developers are
 testing their work in parallel. This is the reason why UAT has higher
 CPU and Memory Allocation.

The limits.cpu and limits.memory are not used anymore in the defaults
and we removed them to align our configurations with Cloud Platform
current defaults/standards.