Reduce Apply namespaces CPU allocation
Runtime CPU usage of each of the Apply-for-legal-aid namespaces stays
well under 100m (usually below 20m).
So allocating 4000m to 20000m for each namespace was an overkill.

The memory currently used by namespace is bellow 3Gi, so setting an
scheduled allocation of 6Gi gives us space for even a replica of all the
containers in the namespace.

We normalized the namespace CPU and Memory allocation for the three
environments/namespaces: UAT, Staging and Production so we have the
same resources everywhere.

The limits.cpu and limits.memory are not used anymore in the defaults
and we removed them to align our configurations with Cloud Platform
current defaults/standards.