Replace Amazon MQ with Amazon ElastiCache for Redis
We never had any success with getting Amazon MQ working with our
combination of Django and Celery.

We tried CloudAMQP for a few months but that developed mysterious
connection problems.

Recently we tried with a freebie Heroku Redis instance and it worked
like a charm.

This change replaces the Amazon MQ instance with a Redis instance
(through Amazon ElastiCache) to finalise the move to Redis.