Reduce resources reserved for monitoring namespace
* Reduce limitrange requests from 125m/250Mi to
10m/100Mi

This change will not have any effect, because all
the pods in the monitoring namespace seem to have
sensible resource values set via the helm chart, 
so these defaults won't be used. But, this change
is worth making in case we add any pods to the
namespace in future, without specifying limits.

* Remove the 'hard limit' on the namespace
* Reduce the request limit from 10000m/24Gi to
500m/12Gi

These changes bring the namespace closer into line
with our defaults for all namespaces, and better
reflect what the monitoring namespace actually
uses.