Additional User Filestore S3 buckets for each environment
In order to support attachments for JSON output, we need new
buckets with a shorter lifecycle policy.

Files will be encrypted with a different key than the existing User
Filestore bucket.
This key will then be passed on to our adapter to fetch and decrypt these files
on request from an external API.  We will use this in combination with
pre-signed URLs.