Reduce limitrange CPU request to 10m from 125m
Our old default value for CPU request per 
container was 125m.

We now know that this is way too much, so this
commit reduces that value down to our current
default of 10m for all namespaces which have it
set to 125m

NB: This commit excludes the formbuilder-product-
page-prod namespace, because the team is in the
process of adjusting resource requirements for all
their namespaces, and I don't want to confuse the
issue.