Reduce limitrange mem. request to 100Mi from 250Mi
Our old default value for memory request per 
container was 250Mi.

We now know that this is too much, so this
commit reduces that value down to our current
default of 100Mi for all namespaces which were
using our default limitrange template.

NB: This commit excludes the formbuilder-product-
page-prod namespace, because the team is in the
process of adjusting resource requirements for all
their namespaces, and I don't want to confuse the
issue.