Reduce CPU default request to 10m for HMPPS
HMPPS namespaces were created when our default
for defaultRequest.cpu was 180m

This is excessive because the containers in these
namespaces generally use around 1m CPU (this is
the case for most namespaces in the cluster) and 
it reduces the CPU available for other workloads.

Once this change is merged, and after the pods
running with the current 180m CPU request have
been cycled, subsequent PRs will reduce the
overall request size for the namespaces, which
will free up these resources.