Add a github actions workflow deleting branches
Merging a PR will trigger the workflow, which will
find the branch the PR relates to, and delete it.
If the user has already deleted the branch, this
workflow will show an error, but nothing terrible
will happen.

It's not strictly necessary to use the tools image
for this. A lighter-weight ruby image would work,
but we'll need the tools image for some workflows,
so it seems simpler to just use it everywhere, 
even if that makes workflows run a little slower
than they could.