Catch errors if branch already deleted
If the user deletes their branch when merging, then the branch
will be missing when this script tries to delete it. This will
raise an error, making it look like something was wrong.

This change swallows the relevant error, so that the action will
complete successfully.

This might sometimes swallow other errors that we would prefer to
see, so the rescue block dumps the error to stdout, so that it's
visible to anyone who views the action log via the github web UI.