Reduce resourcequota for disclosure-checker-prod
The disclosure-checker-production namespace is using 270m CPU and
1200Mi of memory.

This change reduces the CPU request limit from 3000m to 600m, and
the memory request limit from 6000Mi to 5000Mi.

This brings the production namespace into line with the staging
namespace, and frees up 2400m of CPU which is not currently
being used.