Limit new namespaces to 50 pods
We don't set hard limits for CPU and memory anymore. This leaves
us open to a risk that a namespace could spawn unlimited pods,
eventually crashing a worker node.

This change sets a default limit on the number of pods which can
be scheduled into a namespace, which should reduce this risk.