Use dynamodb table to lock environments tf states
This commit depends on https://github.com/ministryofjustice/cloud-platform-concourse/pull/112, which makes the table name available as
an environment variable.

It also depends on https://github.com/ministryofjustice/cloud-platform-concourse/pull/111, which created the dynamodb table.

AFAICT, if the table is missing, or the env var.
is not available, the pipeline would still run
successfully, but would simply not lock the 
terraform states. I might be wrong about this, but
that's what seemed to be happening while I was
figuring this out.