Mount $APPDIR into the tools shell
This enables users to use the tools shell to deploy kubernetes
manifest files from any directory on their local machine.
It requires setting APPDIR before running `make tools-shell`

This is a bit clunky, but if people don't have kubectl installed,
it's the easiest way I could think of to use the tools image for
everything.