Replace CPU & memory request limits with pod count
Instead of limiting the resources that this namespace requests,
limit it to a maximum of 100 pods running concurrently.

This is a lot simpler to understand, and it allows the cluster to
manage resources more efficiently, since requests have the effect
of ring-fencing cluster resources, making them unavailable for
other workloads, whether the namespace is using them or not.