Replace logging request limits with pods: 100
This change removes the cpu and memory request limits from the
logging namespace, and replaces them with a limit of 100 on the
number of pods the namespace may launch.

There are currently 29 pods in the logging namespace, so this
change allows for temporary spikes, e.g. if we were to launch
a full set of new pods before deleting the old ones.