Replace CPU/memory request limits with pod limit
This change removes all the requests.cpu and
requests.memory limits from non-production
namespaces' resourcequotas, and instead sets a 
hard limit on the number of pods each namespace is
allowed to run concurrently.

The values set depend on the number of pods each
namespace is currently running, and is at least
double that number to allow for deployment 
strategies which launch a full set of new pods 
before deleting the old ones (thereby temporarily
doubling the number of running pods).

The precise parameters for choosing the pod limits
are:

```
  case current_pod_count
  when 0..20
    50
  when 21..40
    100
  when 41..60
    120
  else 160
  end
```

This change does not affect any namespaces with
"is-production" set to "true"

Provided this change goes smoothly, a subsequent
PR will apply a similar change for production
namespaces.