Remove the namespace name option from apply script
The concourse deploy pipeline needs to create a service account
in a different namespace from the one belonging to the user.

This change removes the constraint preventing apply from working on
other namespaces.

NB: This is not a good idea to merge into master. We just did this
so that the gitops work could proceed.