Calculate resources_requested from pod data
Now that we no longer have cpu and memory request
limits on namespaces, the namespace-level data
does not contain information on how much resource
is being requested. So, we need to calculate e.g. the total cpu requested by adding up the requests of all containers in all pods in a namespace (ditto for memory).