Add a github action to run bin/spec tests
This action will run all the rspec tests in the bin/spec
directory whenever a PR is raised, or code is pushed to the
master branch.

The only visible output of this action is to put a green tick or
a red cross against the PR, so it may be desirable in future to
add alerting via Slack, or some other, more 'pro-active' result.