Use shorter domain slug
Was `peoplefinder-development-es`

Caused `Error: "domain_name" must start with a lowercase alphabet and
be at least 3 and no more than 28 characters long. Valid characters are
a-z (lowercase letters), 0-9, and - (hyphen).`

Docs say "The domain name of the Elasticsearch cluster to create. This
will be appended with the namespace name and will look like
<team_name>-<environment-name>-<elasticsearch-domain>"

So I think it was trying to make
`peoplefinder-development-peoplefinder-development-es` which blew past
the limit. Using just 'es' in the yaml should hopefully output the name
we wanted in the first place,