Update the env. vars the script needs
The dependency on PIPELINE_CLUSTER is artificial here,but when the script is running in the pipeline, it will be required.

The other env vars are not used. This assumes we'll be copying kubeconfig into the container as per https://github.com/ministryofjustice/cloud-platform-concourse/blob/8e235e57a37283a2b7e5b58fe7b00851f7dcea14/pipelines/live-1/main/build-namespace-changes.yaml