Add GitHub access token for how-out-of-date...
This commit adds a github personal access token to the
how-out-of-date-are-we namespace.

Forthcoming changes to the application include reporting on the
terraform module versions used by namespaces. To do that, the
application needs to be able to call the GitHub API, so it needs a
personal access token.

This token was generated using the cloud-platform-moj GitHub bot user,
so it should be usable indefinitely, regardless of any staff changes.

If you need to recreate the token, note that it does not require any
permissions at all - it just needs to exist. All of our code repos are
public, so no extra permissions are required to get the information the
application needs.