Add Logger class & use it for all command output
This class acts as a single "choke point" for all
command output, so we can locate all of our log
filtering logic there, instead of having to 
duplicate it in the Terraform class.

This will also fix the problem we are currently
having, where the build and plan pipelines report
that all is OK, when in fact there is a problem.

Piping the output of `terraform plan/apply`
through grep has the unfortunate side-effect of
masking the failed exit status of the terraform
command.