Remove the rds-snapshot RDS instance
This instance is no longer required.

It was breaking the build pipeline with this error:

```
Error: error deleting Database Instance
"cloud-platform-a23ffc9abc15d8ba": DBSnapshotAlreadyExists: Cannot
create the snapshot because a snapshot with the identifier
cloud-platform-a23ffc9abc15d8ba-finalsnapshot already exists.
```

I have manually deleted the snapshot and run `terraform apply` to remove
the RDS instance. So this commit brings the code in the repository into
line with the state of AWS and the cluster.