Remove SQS & SNS
This reverts commit 50c038f4 and subsequent tweaks

The SQS/SNS were not being successfully created, 
but a problem with the pipeline was masking the
error.

Now that the problem has been fixed, this code is
breaking our build pipeline. I have notified the
developer, who is working on a fix. In the
meantime, this change allows our pipeline to 
carry on working (so I can find out what else is
broken that we didn't know about).