Add duplicate dev environments to cccd
Due to an urgent, covid-19, legislative change the LAA-Get-Paid team
need to spin up two separate dev environments to fast track
changes.  We do not have time to implement branch builder, so this
will be used to test the entirely new versions of the application
with external and internal users. It is anticipated that they will be
torn down within a month