Sleep for 5 seconds between pods, not 90
There is no point sleeping for a long time between
each pod, when restarting them, because the app.
will be broken from the moment we destroyed the
original elasticache cluster until the pods are
restarted. I'm leaving in a 5 second sleep to give
pods a little time to initialise, but this might
not actually be useful.