Sort lists of files before comparing
This helper method is supposed to check that one directory
matches another, in terms of the names of the files in each directory.
But, the ordering of files returned by `Dir[]` is not guaranteed, so
sometimes the check fails because, e.g.

```
[a, b, c] != [b, a, c]
```

This change sorts each list before comparing, which should prevent these
kinds of false failures.