Include helm-created objects in finder script
The find-deprecated-objects.rb script was omitting objects created by
some helm charts, because they don't have a `last-applied-configuration`

This change iterates over every relevant object in the cluster, so that
any such objects are now included in the list.

NB: This means we now have several *thousand* ReplicaSet objects in the
report, but I would expect those to disappear as the underlying
Deployments are updated.