Detect apiVersion more intelligently
If a kubernetes object is created using the old
api version, and then updated later, the 
"apiVersion" attribute is still the old api version, but the latest-applied-configuration has the new api version.

This change looks for the latter first, and uses that as the relevant api version, if found.

This should result in a more accurate list, without false positives for objects which have already been upgraded.