Ignore ReplicaSet objects
If a deployment, using the old apis, creates a replicaset, then the replicaset is not affected when the deployment is upgraded to the new apis. So there is no point looking for replicasets using the old apis.

Also, no teams should be managing replicasets directly - they should all be using Deployments.