Update deprecated object finder
# Detect apiVersion more intelligently

If a kubernetes object is created using the old api version, and then updated
later, the "apiVersion" attribute is still the old api version, but the
latest-applied-configuration has the new api version.

This change looks for the latter first, and uses that as the relevant api
version, if found.

This should result in a more accurate list, without false positives for objects
which have already been upgraded.

* Ignore ReplicaSet objects

If a deployment, using the old apis, creates a replicaset, then the replicaset
is not affected when the deployment is upgraded to the new apis. So there is no
point looking for replicasets using the old apis.

Also, no teams should be managing replicasets directly - they should all be
using Deployments.