Remove the namespace-usage-report namespace folder
This is a "gitops" namespace, and it is broken since the recent change
to use a terraform module to create our concourse installation.

This change removes the namespace folder altogether (the terraform state
and state lock have already been deleted manually), so that the apply
pipeline can run correctly.

NB: This will cause a problem for the "namespace deleter" pipeline,
which we will have to fix manually.