Update the find-deprecated-objects script
* Stop searching for Ingresses with kubectl

All ingresses show up as using the deprecated APIs
Some of the time this might be valid, but a lot of
the time it's not, and there's no easy way to tell
which is which. So this change simply stops
searching for Ingresses in this way, to avoid a
lot of false alerts.

* Only report kubectl objects when the api version
matches the one in the last-applied-configuration

This should avoid false alerts about helm 
deployments which *have* been updated, but the
object still shows up as using the old API
(because kubernetes sees the update as a noop, 
since the old API is still valid).