Only run the specs when relevant files are changed
The vast majority of the commits and PRs to this repo are to
sub-directories of `namespaces/` For these changes, there is no point
running the specs because they're not relevant.
Github action execution minutes are a limited resource, so we don't want
to waste them. Running unnecessary checks also means we have to wait
longer before we can approve a PR.
This change restricts this github action to only run when *relevant*
changes are proposed/made.