These are the non-prod namespaces which use mod-security. So, these are the ones we need to move to dedicated ingress controllers as a matter of urgency.
This PR will create a separate AWS load-balancer
and ingress controller for each namespace, so that
the teams can add the ingress class annotation to
their ingress definitions to start routing traffic
through the new ingresses.

As soon as all of these namespaces have migrated
their traffic, we will disable mod-security on the
default ingress, which should reduce the problems
we are seeing where the ingress pods are running
out of shared memory.

Related to ministryofjustice/cloud-platform#2224