Add GitHub Action to comment on PR if sufficient
This comment connects to
https://github.com/ministryofjustice/cloud-platform/issues/1931 and
relates to the eventual auto-approval of a PR that is deemed
sufficiently standard.

The action will trigger upon pull request, checkout the users branch,
create a hash of the users newly created namespace and then add a
comment on its suitability. More information on its design can be found
here:
https://github.com/ministryofjustice/cloud-platform-directory-hash/#solution

When the team are comfortable with this solution we can remove the
comment section in this action and uncomment the auto-approval.