Remove SQS queues currently breaking the pipeline
The apply pipeline has been failing over the past 17 hours with the error:
```
Error putting IAM user policy crime-apps-production-create-link-cp-status-job-queue: LimitExceeded: Maximum policy size of 2048 bytes exceeded for user <insert_user>
```
This error is caused by the overuse of a single IAM account. We understand how to fix this problem but it will take some time to fix. For the time-being, to get the apply pipeline working again, this commit will remove the offending SQS queue until a time we can implement a solution.