Remove duplicate namespace resources
All the resources removed in this commit are duplicates of the
[cloud-platform-infrastructure
repository](https://github.com/ministryofjustice/cloud-platform-infrastructure/blob/main/terraform/cloud-platform-components/components.tf).
Any additional changes to the namespace resources should be made to that
repository.

The divergence pipeline will trigger if these are missing so they're not
needed here.