Set force_ssl: true  on cla_backend_rds instances
Since we created the staging and production databases, the default value
of force_ssl has been changed to true, but with an apply_method that is
incompatible with our db engine. Therefore we need to respecify it. This
was stopping the db being created in our training namespace, but we
should specify it in the other two namespaces to avoid confusion