WiP: Start adding Rego policies for namespaces
This change adds a `policy` directory to hold
rego files defining rules all namespaces should
comply with, and a makefile target that tests all
the yaml files in the repo against all of the
policies in that directory.

Initially, this only covers required annotations
and labels for Namespace definitions.

Later changes should add more rules for more
objects, and a github action to apply our policies
to any PRs raised.