Remove hyphenated variables
This change removes the hyphenated (`foo-bar` vs. `foo_bar`) versions of terraform variables in the namespace template.

This probably means that, for new namespaces, we will encounter problems where our modules are expecting hyphenated variables. Spotting these failures in new namespaces is probably the least disruptive way to ensure that we do actually get around to fixing our modules, as per the [ADR]

[ADR]: https://github.com/ministryofjustice/cloud-platform/blob/main/architecture-decision-record/017-Variable-Naming.md#variable-names-in-yaml--terraform-files-are-in-snake-case