Update annotations in all UCPD namespaces
As per conversation in:
https://mojdt.slack.com/archives/C57UPMZLY/p1601638362184400

This PR reassign the business-unit to "HQ" and updates some other outdated or incorrect annotations. Also add the missing `team-name` annotation to all namespaces.

We are aware our team name (from github) is the old one and should be updated too, but that will have to wait as it involves changing a lot more files and re-authenticate to the cluster.