Set minor version to match RDS instances
Removing the minor version from an existing terraform RDS module would
cause terraform to **recreate** the database instance. This is
absolutely not what we want.

This PR sets the minor version to be the same as the deployed RDS
instance. If this works (i.e. if terraform plan reports no changes to
make), then we will have to do this every time an RDS instance
auto-upgrades :(