Pin the version of the "random" provider
Hashicorp upgraded the random provider from 2.30
to 3.0.0 today. This change seems to remove the
`b64` function (there are no release notes, so we
can't confirm this, but stuff breaks).

This PR pins the random provider version to 2.3+
so that namespaces in which the `b64` function is
used continue to plan/apply correctly.