Create new Form Builder service token cache service account
Unfortunately we need to rotate the token used by the old service
account for the service-token-cache app. The way the app utilises it
makes it not possible to simply delete and let K8S regenerate it as that
would mean potential downtime for some users.

We need to create another service account from scratch and move to use
the token generated for that before deleting the old service account.