Update the bin/namespace-reporter.rb script
This outputs the same data as before, but rather than fetching data for
each namespace in turn, using individual `kubectl get ...` and `kubectl
top pod` commands, it now fetches data for all namespaces at once, and
filters out whatever namespaces have been requested via the "pattern"
regular expression.

This means it is a little slower when reporting on a single namespace,
but **much** faster when reporting on multiple namespaces (approx 20
seconds to complete the entire report, on my laptop).